Ls: refactor utility output
